fix(core): close 2 Codex round-7 findings on field-lock + split tests (refs #126)
1. [P2] System fields exempt from lockAllWhen
   `_version` (optimistic-lock token) and other framework-managed
   system fields (id, tenant_id, created_at, updated_at, created_by,
   updated_by) were being matched by `lockAllWhen` like any other
   field. A common UI flow that re-submits `_version` alongside an
   allowlisted edit (e.g., `notes`) on a locked record returned
   `validation.field.locked` instead of reaching the data provider's
   conflict check — masking real concurrency errors AND blocking
   legitimate allowlisted edits. Fix: hard-coded `SYSTEM_FIELD_NAMES`
   set in the checker; lockAllWhen skips them. Per-field `lockWhen`
   declared explicitly on a system field still applies (intentional
   authorial override).

2. [P2] Translatable fields normalized for same-value comparison
   When the provider stores i18n content as a locale map
   (`{ en: "Hello", zh: "你好" }`) and the client submits a plain
   string, structural equality saw them as different shapes and
   rejected unchanged re-submits as immutable violations. Fix: when
   the field is `translatable: true`, the existing value is a plain-
   prototype object, and the new value is a string, treat the string
   as unchanged if it equals any locale value in the existing map.
   Falls back to structural equality otherwise.

## Test-file split
field-lock-handler.test.ts grew past 500 with the round-6/7
additions. Split:
- `field-lock-handler.test.ts` (290) — round-5: ctx.update wrapper
  blocks immutable / lockWhen, subset-write scoping, Date/string.
- `field-lock-classify.test.ts` (250, NEW) — round-6/7: cross-entity
  scoping, deterministic mixed-violation classification, translatable
  same-value normalization.

Suite: 4452 pass / 3 skip / 0 fail.
